The edge reconfigure manager is in its own file.
